requirements.txt
openai
elevenlabs
pyaudio
pygame
faster_whisper
torch
python-dotenv